exobrain: update
Synchronize exobrain content with authoritative copy.
Also switch inclusion mechanism from Git submodule to simple copy&paste.

Extracting the exobrain to a different workspace has been a great idea,
using a submodule to relink it to this repo a dumb one.
The exobrain isn't something inherently versioned, only the latest
version matters at all. So using Git for this job brings nothing but
worthless mental overhead.
I still had to manually sync the submodule in this repo to the exobrain
one, that in turn had to be _quite artificially_ be synced to the
authoritative copy.

Directly rsyncing the local copy from the authoritative one is a much
simpler and effective method to manage this stuff.